refactor - tested all util files.
