+++ util/dots.py
@@ -1,0 +1,4 @@
+from numpy.linalg import multi_dot
+
+def dots(*arg):
+    return multi_dot(arg)
